requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.4
django-debug-toolbar==3.2
djangorestframework==3.12.2
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
python-decouple==3.3
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
